Add public `DictionaryBuilder` APIs
